requirements.txt
# Project Sovereign - Dependencias Core

# Procesamiento de video
yt-dlp>=2023.10.13
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# IA y transcripción
openai-whisper>=20231117
google-generativeai>=0.3.0

# Utilidades
textdistance>=4.5.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Base de datos (ya instalado)
psycopg2-binary>=2.9.9

# Procesamiento de imagen/video
pillow>=10.1.0

# Audio processing
pyloudnorm>=0.1.1

# Motor B full pipeline
edge-tts>=6.1.12
requests>=2.31.0
pytrends>=4.9.2
feedparser>=6.0.11
